Implement webpack manifest for JavaScript bundle management
A new feature has been added to manage JavaScript bundles using a webpack manifest plugin, updating overall webpack configuration and updating the package.json. Adjustments also include validation for authorisation in the DisplayController, and updates to the date shown in the bundle-report.html file. A condition has been added in the dashboard/default.php to check for unset or empty JS bundle and a comment added in case the condition is met.