Refactored UsersExport extension adding search and date range filters
The UsersExport extension was refactored to allow search functionality and date range filters on the front end. The search and range filters are also implemented on the back end in UsersController.php and UsersModel.php. It also includes the addition of useful tools such as lodash and csvUtils for data manipulation and CSV data export functionality.