Add user authentication and permissions checks
New checks have been introduced to ensure that users are authenticated and hold the necessary permissions to access administrative areas in the UsersController.php file. In addition, adjustments have been made for better fetching and exporting of user data. Permissions fields have been added to usersexport.xml and updates done in the UI for clarity and improved user experience.