Refactor UsersExport module and update authorization control
The commit renames the file 'installer_script.php' to 'InstallerScript.php' and updates its file path, improving the script organization. It also redefines the admin path and modifies the procedure to create a log marker file. The permissions section from 'usersexport.xml' file was removed. In 'DisplayController.php', the check for authorization was simplified by only focusing on guest users, improving the readability and the efficiency of the code.